hydraulic-app/requirements.txt
kafka-python-ng==2.2.3
prometheus-client==0.19.0
requests==2.31.0
pymongo==4.6.0
pyspark==3.5.0
mlflow==2.9.2
